[PORT FROM MAIN] Fix StackOverflowError exception in media provider
BZ: 60211

When the USB disk has been removed brutaly, the mDirectoryCache is cleared and the full
path of the contents is rebuilt recursively causing a StackOverflowError if the path
to build is containing too much elements.

We catch the StackOverflowError exception here as a workaround for this issue.

Category: aosp improvement
Domain: AOSP-Framework-Media
Severity: medium
Origin: internal
Upstream-Candidate: yes

Signed-off-by: lwan89x <liangx.wang@intel.com>

Change-Id: I576ed9b8d5d03593b55bd255409c41a8137fceac
Orig-Change-Id: I967896c621dae9358fa11af2b8378503aaf078cd