[PORT FROM R4.1]RingtonePickerActivity: Fix error of attempting to access a deactivated CursorWindow.
BZ: 76836

There is a race in RingtonePickerActivity on some condition: if choose
a ringtone using hard key, it will post the playing ringtone runnable
in 300ms' delay. Before this runnable running, if OK or Cancel button
was pressed, mCursor will be deactivated. Then java crash will happen
in previous' runnable due to access to a deactivated cursor. This error
always happened in monkey test.

We add a cursor deactivated flag to resolve this race condition.

Change-Id: If7fc796b444c761e9c0c933365848fa0b50b96f7
Orig-Change-Id: I0d24e79bf5aa393b7cfd58d126b911ee8d520cf3
Signed-off-by: lwan89x <liangx.wang@intel.com>
Reviewed-on: http://android.intel.com:8080/84389
Reviewed-by: Yin, Fengwei <fengwei.yin@intel.com>
Reviewed-by: Li, XiaojingX <xiaojingx.li@intel.com>
Tested-by: Li, XiaojingX <xiaojingx.li@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>